fetch datasource query settings from query editor
